config: update default param options
Prior to this commit, we set a height param that would result in
an aspect ratio of 1:1. Now, we specify a width and let the aspect
ratio determine the correct height.

We also specified cropping params, but those have been removed as
well. Typically, you wouldn't want to crop e-commerce images like
that. Now, those params have been removed.